Simplify the engine context
This commit refactors most of how the WASI execution engine is set up.
Specifically, when pre-instantiating, an empty store is created, and only
when creating a concrete instance for execution, a store with actual
state is created and used to instantiate the module.
